requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
google-cloud-aiplatform==1.56.0
werkzeug==3.0.3
gunicorn==22.0.0
google-cloud-storage==2.14.0
